Make the enemies spawn on the whole map
