+++ Assets/Prefabs/EnemyMul.prefab
@@ -192,7 +192,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 950bb97783369c2469af1ccb527f1609, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  BulletPrefab: {fileID: 100000, guid: 203c159f6f40ab8cdb9fc152a38998ef, type: 3}
+  BulletPrefab: {fileID: 100000, guid: 0e25dcf5567af14bba3a30173a2e940e, type: 3}
   bulletForce: 5
   fireRate: 3
   health:
