+++ Assets/Scenes/Scene1.unity
@@ -156,7 +156,7 @@ MonoBehaviour:
   PooledPrefabsList:
   - Prefab: {fileID: 100000, guid: 26e8740d898c8544b984097438d483c0, type: 3}
     PrewarmCount: 0
-  - Prefab: {fileID: 100000, guid: 203c159f6f40ab8cdb9fc152a38998ef, type: 3}
+  - Prefab: {fileID: 100000, guid: 0e25dcf5567af14bba3a30173a2e940e, type: 3}
     PrewarmCount: 0
   - Prefab: {fileID: 4933005563524840836, guid: dd39defb290aa054d9947b9a64dc89c3, type: 3}
     PrewarmCount: 0
@@ -315,7 +315,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Texture: {fileID: 8400000, guid: 750594b4d7d6f4817bc8027159c1d75f, type: 2}
+  m_Texture: {fileID: 8400000, guid: 816f625d6574b4c9f83bb30f19b08cf5, type: 2}
   m_UVRect:
     serializedVersion: 2
     x: 0
@@ -568,7 +568,7 @@ MonoBehaviour:
       SourceHashToOverride: 0
       OverridingTargetPrefab: {fileID: 0}
     - Override: 0
-      Prefab: {fileID: 100000, guid: 203c159f6f40ab8cdb9fc152a38998ef, type: 3}
+      Prefab: {fileID: 100000, guid: 0e25dcf5567af14bba3a30173a2e940e, type: 3}
       SourcePrefabToOverride: {fileID: 0}
       SourceHashToOverride: 0
       OverridingTargetPrefab: {fileID: 0}
@@ -714,11 +714,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 2027122270}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 58127b4fbee90f79699f2a483a4c6299, type: 3}
+  m_Script: {fileID: 11500000, guid: d73e3b61fc2d16f4fbf932d28e0dd802, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  spawnDelay: 1
+  spawnTime: 3
+  EnemyPrefab: {fileID: 4933005563524840836, guid: dd39defb290aa054d9947b9a64dc89c3, type: 3}
   initialServerSeed: 7312
-  ObstaclePrefab: {fileID: 5019422062460965281, guid: a3d17b98e4522319485f323529ff9d1e, type: 3}
+  ObstaclePrefab: {fileID: 5019422062460965281, guid: 640acbdb0513b0b75b0538657fd48094, type: 3}
   roomSize: 21
   doorSize: 5
   roomsInRow: 7
@@ -804,7 +807,7 @@ Camera:
     serializedVersion: 2
     m_Bits: 4294967295
   m_RenderingPath: -1
-  m_TargetTexture: {fileID: 8400000, guid: 750594b4d7d6f4817bc8027159c1d75f, type: 2}
+  m_TargetTexture: {fileID: 8400000, guid: 816f625d6574b4c9f83bb30f19b08cf5, type: 2}
   m_TargetDisplay: 0
   m_TargetEye: 0
   m_HDR: 1
