--- Assets/Scripts/EnemySpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9bd0f62d181a8c842a1963483319782f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
